End to end benchmark

+++ include/orderbook/Order.h
@@ -7,8 +7,8 @@
 #include <vector>
 
 using OrderId  = std::uint64_t;
-using Price    = std::int64_t;
-using Quantity = std::uint64_t;
+using Price    = std::uint32_t;
+using Quantity = std::uint32_t;
 using OrderIds = std::vector<OrderId>;
 
 enum class OrderType { GoodTillCancel, FillAndKill,
